@@ -9,16 +9,16 @@
     "@condict/rich-text-editor": "*",
     "@condict/table-editor": "*",
     "@condict/ui": "*",
-    "@types/react": "^18.0.15",
-    "@types/react-dom": "^18.0.6",
-    "next": "^12.2.3"
+    "@types/react": "^18.0.25",
+    "@types/react-dom": "^18.0.9",
+    "next": "^12.3.4"
   },
   "dependencies": {
-    "babel-plugin-styled-components": "^2.0.2",
-    "immer": "^9.0.15",
+    "babel-plugin-styled-components": "^2.0.7",
+    "immer": "^9.0.16",
     "platform": "^1.3.6",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
-    "styled-components": "5.1.1"
+    "styled-components": "^5.3.6"
   }
 }
