@@ -8,13 +8,13 @@
     "setup": "npm install && npm run bootstrap && npm run build",
     "bootstrap": "lerna bootstrap --hoist",
     "clean": "rimraf packages/*/dist",
-    "build": "node scripts/build -g build-tools && node scripts/build -g utils && node scripts/build -g server && node scripts/build -g ui && npm run -s build:app",
+    "build": "node scripts/build.mjs -g build-tools && node scripts/build.mjs -g utils && node scripts/build.mjs -g server && node scripts/build.mjs -g ui && npm run -s build:app",
     "build:children": "lerna run build",
     "build:app": "cd packages/app && npm run -s build",
-    "compile": "rollup --config config/rollup.config.js",
-    "watch:utils": "node scripts/build -w -g utils",
-    "watch:ui": "node scripts/build -w -g ui",
-    "watch:server": "node scripts/build -w -g server",
+    "compile": "rollup --config config/rollup.config.mjs",
+    "watch:utils": "node scripts/build.mjs -w -g utils",
+    "watch:ui": "node scripts/build.mjs -w -g ui",
+    "watch:server": "node scripts/build.mjs -w -g server",
     "watch:server-graphql": "cd packages/server && npm run -s watch:graphqlTypes",
     "watch:app": "cd packages/app && npm run -s watch",
     "serve": "cd ./dev && npx --no-install next",
@@ -25,24 +25,24 @@
     "test": "lerna run test"
   },
   "devDependencies": {
-    "@rollup/plugin-commonjs": "^22.0.1",
-    "@rollup/plugin-json": "^4.1.0",
-    "@rollup/plugin-node-resolve": "^13.3.0",
-    "@rollup/plugin-replace": "^4.0.0",
-    "@rollup/plugin-typescript": "^8.3.3",
-    "@typescript-eslint/eslint-plugin": "^5.30.7",
-    "@typescript-eslint/parser": "^5.30.7",
+    "@rollup/plugin-commonjs": "^23.0.3",
+    "@rollup/plugin-json": "^5.0.2",
+    "@rollup/plugin-node-resolve": "^15.0.1",
+    "@rollup/plugin-replace": "^5.0.1",
+    "@rollup/plugin-typescript": "^10.0.1",
+    "@typescript-eslint/eslint-plugin": "^5.45.0",
+    "@typescript-eslint/parser": "^5.45.0",
     "chalk": "^4.1.2",
     "command-line-args": "^5.2.1",
     "cross-env": "^7.0.3",
-    "eslint": "^8.20.0",
+    "eslint": "^8.28.0",
     "eslint-plugin-jsx-a11y": "^6.6.1",
-    "eslint-plugin-react": "^7.30.1",
-    "lerna": "^5.2.0",
+    "eslint-plugin-react": "^7.31.11",
+    "lerna": "^6.1.0",
     "npm-run-all": "^4.1.5",
     "rimraf": "^3.0.2",
-    "rollup": "^2.77.0",
-    "tslib": "^2.4.0",
-    "typescript": "^4.7.4"
+    "rollup": "^3.5.1",
+    "tslib": "^2.4.1",
+    "typescript": "^4.9.3"
   }
 }
