@@ -26,9 +26,9 @@
     "@condict/platform": "^0.1.0",
     "@condict/server": "^0.1.0",
     "chokidar": "^3.5.3",
-    "graphql": "^16.5.0",
+    "graphql": "^16.6.0",
     "node-fetch": "^2.6.7",
-    "ws": "^8.8.1"
+    "ws": "^8.11.0"
   },
   "optionalDependencies": {
     "fsevents": "^2.3.2"
@@ -39,27 +39,27 @@
     "@condict/rich-text-editor": "^0.1.0",
     "@condict/table-editor": "^0.1.0",
     "@condict/ui": "^0.1.0",
+    "@electron/rebuild": "^3.2.10",
     "@fluent/bundle": "^0.17.1",
     "@fluent/react": "^0.14.1",
-    "@types/node": "^18.6.1",
+    "@types/node": "^18.11.10",
     "@types/node-fetch": "^2.6.2",
-    "@types/react": "^18.0.15",
-    "@types/react-dom": "^18.0.6",
+    "@types/react": "^18.0.25",
+    "@types/react-dom": "^18.0.9",
     "@types/shallowequal": "^1.1.1",
-    "@types/styled-components": "^5.1.25",
+    "@types/styled-components": "^5.1.26",
     "cross-env": "^7.0.3",
-    "electron": "^19.0.9",
-    "electron-builder": "^23.3.3",
-    "electron-rebuild": "^3.2.8",
-    "immer": "^9.0.15",
+    "electron": "^22.0.0",
+    "electron-builder": "^23.6.0",
+    "immer": "^9.0.16",
     "mdi-react": "^9.0.0",
     "nanoid": "^3.3.4",
-    "node-abi": "^3.22.0",
+    "node-abi": "^3.30.0",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-is": "^17.0.2",
     "rimraf": "^3.0.2",
     "shallowequal": "^1.1.0",
-    "styled-components": "5.1.1"
+    "styled-components": "^5.3.6"
   }
 }
